app/requirements.txt
numpy
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-jaeger
httpx